fix(cli/volumes): move `s3.rs` to its `mod.rs`
fix(cli/volumes): move `s3.rs` to its `mod.rs`
